Added stroke fonts.
git-svn-id: http://svn.code.sf.net/p/freeglut/code/trunk/freeglut/freeglut@25 7f0cb862-5218-0410-a997-914c9d46530a

+++ freeglut/include/GL/freeglut.h
@@ -447,7 +447,7 @@ FGAPI int     FGAPIENTRY glutStrokeWidth( void* font, int character );
 FGAPI int     FGAPIENTRY glutBitmapLength( void* font, const char* string );
 FGAPI int     FGAPIENTRY glutStrokeLength( void* font, const char* string );
 FGAPI int     FGAPIENTRY glutBitmapHeight( void* font );
-FGAPI int     FGAPIENTRY glutStrokeHeight( void* font );
+FGAPI GLfloat FGAPIENTRY glutStrokeHeight( void* font );
 FGAPI void    FGAPIENTRY glutBitmapString( void* font, const char *string );
 FGAPI void    FGAPIENTRY glutStrokeString( void* font, const char *string );
